Add major refactor for domupdates and eventlisteners
